% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel.R
\name{pbGetChannelInfo}
\alias{pbGetChannelInfo}
\alias{pbGetChannelInfo.default}
\alias{print.pbChannelInfo}
\alias{summary.pbChannelInfo}
\title{Details for a channel}
\usage{
pbGetChannelInfo(channel, no_recent_pushes = FALSE)

\method{pbGetChannelInfo}{default}(channel, no_recent_pushes = FALSE)

\method{print}{pbChannelInfo}(x, ...)

\method{summary}{pbChannelInfo}(object, ...)
}
\arguments{
\item{channel}{The name of a Pushbullet channel as a string}

\item{no_recent_pushes}{Should the returned returned object exclude recent pushs?
\code{FALSE} (the default) will return up to 10 pushes. \code{TRUE} will exclude them}

\item{x}{Default object for \code{print} method}

\item{...}{Other optional arguments}

\item{object}{Default object for \code{summary} method}
}
\value{
a list with infoabout a channel
}
\description{
Details for a channel
}
\examples{
xkcd <- pbGetChannelInfo("xkcd", TRUE)
summary(xkcd)
}
