% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pushes.R
\name{pbGetPost}
\alias{pbGetPost}
\title{Get messages posted via Pushbullet}
\usage{
pbGetPost(apikey = .getKey(), limit = 10, verbose = FALSE,
  debug = FALSE)
}
\arguments{
\item{apikey}{The API key used to access the service. It can be
supplied as an argument here, via the global option
\code{rpushbullet.key}, or via the file \code{~/.rpushbullet.json}
which is read at package initialization (and, if found, also sets
the global option).
\code{~/.rpushbullet.json} which is read at package
initialization.}

\item{limit}{Limit number of post. Default is 10.}

\item{verbose}{Boolean switch to add additional output}

\item{debug}{Boolean switch to add even more debugging output}
}
\value{
A data.frame result record is return
}
\description{
This function get messages posted to Pushbullet.
}
\examples{
\dontrun{
pbGetPost()
}
}
\author{
Chanyub Park
}
