% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pushes.R
\name{pbGetPosts}
\alias{pbGetPosts}
\title{Get messages posted via Pushbullet}
\usage{
pbGetPosts(apikey = .getKey(), limit = 10)
}
\arguments{
\item{apikey}{The API key used to access the service. It can be
supplied as an argument here, via the global option
\code{rpushbullet.key}, or via the file \code{~/.rpushbullet.json}
which is read at package initialization (and, if found, also sets
the global option).
\code{~/.rpushbullet.json} which is read at package
initialization.}

\item{limit}{Limit number of post. Default is 10.}
}
\value{
A data.frame result record is returned
}
\description{
This function gets messages posted to Pushbullet.
}
\examples{
\dontrun{
pbGetPosts()
}
}
\author{
Chanyub Park
}
