% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{pbGetUser}
\alias{pbGetUser}
\alias{pbGetUser.default}
\alias{print.pbUser}
\alias{summary.pbUser}
\title{Get info about a user}
\usage{
pbGetUser(apikey = .getKey())

\method{pbGetUser}{default}(apikey = .getKey())

\method{print}{pbUser}(x, ...)

\method{summary}{pbUser}(object, ...)
}
\arguments{
\item{apikey}{The API key used to access the service. It can be
supplied as an argument here, or via the file
\code{~/.rpushbullet.json} which is read at package
initialization.}

\item{x}{Default object for \code{print} method}

\item{...}{Other optional arguments}

\item{object}{Default object for \code{summary} method}
}
\value{
Invisibly returns info about a user
}
\description{
Get info about a user
}
\examples{
\dontrun{
me <- pbGetUser()
summary(me)
}
}
