% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pbLoadConfig}
\alias{pbLoadConfig}
\title{Load the pushbullet account information from a file.}
\usage{
pbLoadConfig(file, text, use = TRUE)
}
\arguments{
\item{file}{File to load the information from.}

\item{text}{If 'file' is not supplied and this is a character
string, then data are read from the value of 'text' via a text
connection.}

\item{use}{Whether to save the read list in the environment for
use (TRUE) or just return what was read (FALSE, more for
debugging).}
}
\value{
The list of apikey, devices, and defaultdevice, invisibly.
}
\description{
Load the pushbullet account information from a file.
}

