% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pbSaveConfig}
\alias{pbSaveConfig}
\title{Save the pushbullet account information to a file.}
\usage{
pbSaveConfig(apikey, devices, defaultdevice, file = "", overwrite = FALSE)
}
\arguments{
\item{apikey}{The API key used to access the service.}

\item{devices}{The devices stored as potential recipients.}

\item{defaultdevice}{The default recipient device for posts. See `pbPost` for details.}

\item{file}{File to save the information to.}

\item{overwrite}{boolean. Whether overwrite the file if it already exists; if TRUE, all data in the file is lost}
}
\value{
The list of apikey, devices, and defaultdevice, invisibly.
}
\description{
The default behavior of other rpushbullet functions relies solely
on the apikey and device list being accessible via
\sQuote{options()}. This information can be cached in a local file
and recalled when the library is loaded.
}
\details{
The default file is \sQuote{~/.rpushbullet.json}, though any file
can be provided; note that the default filename is opened when the
library is attached, otherwise you'll need to manually load it
with \sQuote{pbLoadConfig()}.

If file is not provided then the string that would be saved to it
is printed to the console.
}

