% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{pbValidateConf}
\alias{pbValidateConf}
\title{Check if a configuration is valid}
\usage{
pbValidateConf(conf = NULL)
}
\arguments{
\item{conf}{Either a file path (like \code{~/.rpushbullet.json}) or a JSON string.
If \code{NULL} (the default), the value of \code{getOption("rpushbullet.dotfile")}
will be used.}
}
\value{
\code{TRUE} if both the api key and \emph{all} devices are vaild. \code{FALSE} otherwise.
}
\description{
Check if a configuration is valid
}
\examples{
pbValidateConf('{"key":"a_fake_key","devices":["dev_iden1","dev_iden2"]}')

}
