\name{RPushbullet-package}
\alias{RPushbullet-package}
\alias{RPushbullet}
\docType{package}
\title{
  R interface to the Pushbullet service
}
\description{
  The \href{http://www.pushbullet.com}{Pushbullet} service
  permits users to pass messenges between their computers, phones and
  other devices such as tablets.  It offers immediacy
  which is perfect for alerting, and much more.

  This package provides a programmatic interface from R.
}
\details{
  The \href{https://www.pushbullet.com/api}{Pushbullet API} offers a
  RESTful interface. It requires an API key which can be obtained free
  of charge.  Given the API key, and one or more registered devices,
  users can push message to one or more device.
}
\author{
  Dirk Eddelbuettel 
}
\references{
  See the Pushbullet documentation at the
  \href{http://www.pushbullet.com}{site}.
}
\keyword{package}
