% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pbGetDevices}
\alias{pbGetDevices}
\title{Get registered Pushbullet devices}
\usage{
pbGetDevices(apikey)
}
\arguments{
\item{apikey}{The API key used to access the service. It can be
supplied as an argument here, or via the file
\code{~/.rpushbullet.json} which is read at package
initialization.}
}
\value{
A JSON result record is return invisibly
}
\description{
Retrieve the list of devices registered for the given API key.
}
\details{
This function invokes the \sQuote{devices} functionality of
the Pushbullet API; see \url{https://docs.pushbullet.com/v2/devices} for more
details.
}
\author{
Dirk Eddelbuettel
}

